Added Password Digits command

+++ sqls/db_structure_request.sql
@@ -1,8 +1,8 @@
 drop table if exists bot_helper31_request;
 create table bot_helper31_request (
-    id int generated always as identity primary key,
+    id bigint generated always as identity primary key,
     requested_on timestamp with time zone not null default current_timestamp,
-    sender_id int,
+    sender_id bigint,
     user_name text,
     last_name text,
     first_name text,
@@ -11,7 +11,7 @@ create table bot_helper31_request (
 );
 drop table if exists bot_helper31_password_settings;
 create table bot_helper31_password_settings (
-    chat_id int not null primary key,
+    chat_id bigint not null primary key,
     length int not null default 12,
     use_upper_case int not null default 1,
     use_lower_case int not null default 1,
